Confirm gift request for organization
