+++ app/mailers/user_mailer.rb
@@ -21,11 +21,6 @@ class UserMailer < Notifier
     mail to: @user.email, subject: I18n.t('mailer.donation.canceled')
   end
 
-  def confirm_gift_request_email_to_org(user, organization, gift_item)
-    gift_request_email_elements(user, organization, gift_item)
-    mail to: @organization.email, subject: I18n.t('mailer.gift_request.confirmed')
-  end
-
   def cancel_gift_request_email_to_org(user, organization, gift_item)
     gift_request_email_elements(user, organization, gift_item)
     mail to: @organization.email, subject: I18n.t('mailer.gift_request.canceled')
