+++ app/views/organization_mailer/cancel_gift_request_email_to_org.html.erb
@@ -7,6 +7,7 @@
     <p>
       <%= t('mailer.hello') %> <%= @organization.name %>
         <%= @user.full_name %> <%= t('mailer.canceled') %> <%= t('mailer.the_m') %> <%= t('mailer.gift_request') %> <%= t('mailer.for') %> <%= link_to @gift_item.title, user_gift_item_path(@gift_item.user.id, @gift_item.id) %>.
+        <%= t('mailer.reason') %>: <%= @reason %>
         <%= t('mailer.you_can_contact') %> <%= @user.full_name %> <%= t('mailer.writing_to') %> <%= @user.email %>
       <p><%= t('mailer.have_great_day')%></p>
     </p>
