--- app/views/user_mailer/confirm_gift_request_email_to_org.html.erb
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta content='text/html; charset=UTF-8' http-equiv='Content-Type' />
-  </head>
-  <body>
-    <p>
-      <%= t('mailer.hello') %> <%= @organization.name %>
-        <%= @user.full_name %> <%= t('mailer.confirmed') %> <%= t('mailer.the_m') %> <%= t('mailer.gift_request') %> <%= t('mailer.for') %> <%= link_to @gift_item.title, user_gift_item_path(@gift_item.user.id, @gift_item.id) %>.
-        <%= t('mailer.you_can_contact') %> <%= @user.full_name %> <%= t('mailer.writing_to') %> <%= @user.email %>
-      <p> <%= t('mailer.have_great_day')%></p>
-    </p>
-  </body>
-</html>
