+++ spec/controllers/organization_gift_requests_controller_spec.rb
@@ -149,4 +149,139 @@ describe OrganizationGiftRequestsController do
       end
     end
   end
+
+  describe '#confirm' do
+    context 'When logged' do
+      before(:each) do
+        sign_in user
+        user.reload
+      end
+      context 'When the organization owns the gift request' do
+        let!(:gift_request) do
+          create :gift_request, organization: organization, user: user, gift_item: gift_item
+        end
+        context 'When confirming a gift_request' do
+          context 'When gift_request done is false' do
+            it 'changes done' do
+              expect { post :confirm, id: gift_request.id }.to change { gift_request.reload.done }
+            end
+            it 'changes done to true' do
+              post :confirm, id: gift_request.id
+              gift_request.reload
+              expect(gift_request.done).to be true
+            end
+            it 'redirects' do
+              post :confirm, id: gift_request.id
+              expect(response.status).to eq 302
+            end
+            it 'sends an email' do
+              (expect do
+                post :confirm, id: gift_request.id
+              end).to change { ActionMailer::Base.deliveries.count }.by(1)
+            end
+          end
+          context 'When gift_request done is true' do
+            before(:each) do
+              gift_request.update_attributes(done: true)
+            end
+            it 'responds status forbidden' do
+              post :confirm, id: gift_request.id
+              expect(response.status).to eq 403
+            end
+            it 'does not send an email' do
+              (expect do
+                post :confirm, id: gift_request.id
+              end).not_to change { ActionMailer::Base.deliveries.count }
+            end
+          end
+          context 'When nothing to give left' do
+            before(:each) do
+              gift_item.update_attributes(given: gift_item.quantity)
+            end
+            it 'responds status forbidden' do
+              post :confirm, id: gift_request.id
+              expect(response.status).to eq 403
+            end
+            it 'does not change done status' do
+              expect { post :confirm, id: gift_request.id }
+                .not_to change { gift_request.reload.done }
+            end
+            it 'does not change given of gift_item' do
+              expect { post :confirm, id: gift_request.id }
+                .not_to change { gift_item.reload.given }
+            end
+            it 'does not send an email' do
+              (expect do
+                post :confirm, id: gift_request.id
+              end).not_to change { ActionMailer::Base.deliveries.count }
+            end
+          end
+          context 'When asking more than what is left' do
+            before(:each) do
+              gift_request.update_attributes(quantity: 6)
+            end
+            context 'When having something to give' do
+              before(:each) do
+                gift_item.update_attributes(given: 5, quantity: 10)
+              end
+              it 'increments given to quantity' do
+                post :confirm, id: gift_request.id
+                gift_item.reload
+                expect(gift_item.given).to eq gift_item.quantity
+              end
+            end
+            context 'When asking less that what is left' do
+              before(:each) do
+                gift_request.update_attributes(quantity: 5)
+              end
+              it 'increments given gift_request.quantity' do
+                expect { post :confirm, id: gift_request.id }
+                  .to change { gift_item.reload.given }.by gift_request.reload.quantity
+              end
+            end
+          end
+        end
+      end
+      context 'When organization does not own the gift request' do
+        let!(:gift_request) do
+          create :gift_request, organization: organization2, user: user, gift_item: gift_item
+        end
+        context 'When trying to confirm a request' do
+          it 'responds status forbidden' do
+            post :confirm, id: gift_request.id
+            expect(response.status).to eq 403
+          end
+          it 'does not change done status' do
+            expect { post :confirm, id: gift_request.id }
+              .not_to change { gift_request.reload.done }
+          end
+          it 'does not change given of gift_item' do
+            expect { post :confirm, id: gift_request.id }.not_to change { gift_item.reload.given }
+          end
+          it 'does not send an email' do
+            (expect do
+              post :confirm, id: gift_request.id
+            end).not_to change { ActionMailer::Base.deliveries.count }
+          end
+        end
+      end
+    end
+    context 'When user is unlogged' do
+      before :each do
+        gift_item.update_attributes(user: user)
+      end
+      let!(:gift_request) do
+        create :gift_request, organization: organization, user: user, gift_item: gift_item
+      end
+      it 'should render login page' do
+        post :confirm, id: gift_request.id
+        expect(response).to redirect_to '/users/sign_in'
+      end
+      it 'does not send an email' do
+        (expect do
+          post :confirm, id: gift_request.id
+        end).not_to change { ActionMailer::Base.deliveries.count }
+      end
+    end
+  end
 end
